Setup the form submit interception function.
A couple of notes... I hooked up the onsubmit function in the javascript
file globally to avoid loading the js first. Then, I needed to avoid a post-
back and reload of the page (I guess not, but it makes a little more sense).
And, now, I can intercept the form values, do the work, and reset before
returning false.